refactor: streamline capture handling and improve Wayland support
- Removed advisory target handling for Wayland portal in recording.rs, simplifying the capture target resolution process.
- Updated system.rs to enhance the capture capabilities matrix for macOS and Windows, reflecting changes in backend support.
- Introduced a new function `pack_rows` in encoder/mod.rs to handle row packing for FFmpeg's rawvideo demuxer, ensuring proper stride handling.
- Added `adopt_source_size` in recording/mod.rs to adjust capture target dimensions based on actual source delivery, particularly for Wayland.
- Modified pipeline.rs to accept a `CaptureSource` trait object, allowing for more flexible capture source management.
- Enhanced camera handling in recording/mod.rs to ensure a single thread manages the camera, improving frame delivery to both preview and recording.
- Updated documentation in architecture files to reflect changes in camera handling and capture pipeline structure.
- Introduced readback rate configuration in capturekit to optimize frame readback for Windows Graphics Capture.

+++ apps/desktop/src-tauri/src/cli.rs
@@ -693,9 +693,9 @@ enum RecAction {
 #[derive(Subcommand)]
 enum SelectAction {
     /// Record a display by id (from `displays list`).
-    Screen { id: u32 },
+    Screen { id: u64 },
     /// Record a window by id (from `windows list`).
-    Window { id: u32 },
+    Window { id: u64 },
     /// Record a region given as X,Y,W,H in physical pixels.
     Region {
         #[arg(value_name = "X,Y,W,H")]
@@ -759,13 +759,13 @@ enum ProfileAction {
 enum ScreenshotTarget {
     /// A whole display by id (from `displays list`).
     Display {
-        id: u32,
+        id: u64,
         #[command(flatten)]
         shot: ShotArgs,
     },
     /// One application window by id (from `windows list`).
     Window {
-        id: u32,
+        id: u64,
         #[command(flatten)]
         shot: ShotArgs,
     },
